done recommend for home and detail

+++ lib/objects/voucher_children/limited_amount_voucher_without_end_time.dart
@@ -45,7 +45,7 @@ class LimitedAmountVoucherWithoutEndTime extends Voucher
     String? voucherName,
     DateTime? startTime,
     double? discountValue,
-    double? minimumPurchase,
+    int? minimumPurchase,
     int? maxUsagePerPerson,
     bool? isVisible,
     bool? isEnabled,
