+++ lib/objects/voucher_children/limited_percentage_voucher_with_end_time.dart
@@ -12,7 +12,7 @@ class LimitedPercentageVoucherWithEndTime
     implements LimitedInterface, EndTimeInterface, PercentageInterface {
   int _maximumUsage;
   int _usageLeft;
-  double _maximumDiscountValue;
+  int _maximumDiscountValue;
   DateTime _endTime;
 
   LimitedPercentageVoucherWithEndTime({
@@ -34,7 +34,7 @@ class LimitedPercentageVoucherWithEndTime
     required int maximumUsage,
     required int usageLeft,
     required DateTime endTime,
-    required double maximumDiscountValue,
+    required int maximumDiscountValue,
   }) :
         _maximumUsage = maximumUsage,
         _usageLeft = usageLeft,
@@ -57,9 +57,9 @@ class LimitedPercentageVoucherWithEndTime
   set endTime(DateTime value) => _endTime = value;
 
   @override
-  double get maximumDiscountValue => _maximumDiscountValue;
+  int get maximumDiscountValue => _maximumDiscountValue;
   @override
-  set maximumDiscountValue(double value) => _maximumDiscountValue = value;
+  set maximumDiscountValue(int value) => _maximumDiscountValue = value;
 
   @override
   void updateVoucher({
@@ -67,7 +67,7 @@ class LimitedPercentageVoucherWithEndTime
     String? voucherName,
     DateTime? startTime,
     double? discountValue,
-    double? minimumPurchase,
+    int? minimumPurchase,
     int? maxUsagePerPerson,
     bool? isVisible,
     bool? isEnabled,
@@ -77,7 +77,7 @@ class LimitedPercentageVoucherWithEndTime
     int? maximumUsage,
     int? usageLeft,
     DateTime? endTime,
-    double? maximumDiscountValue,
+    int? maximumDiscountValue,
   }) {
     super.updateVoucher(
       voucherID: voucherID,
